requirements.txt
fastapi
uvicorn[standard]
line-bot-sdk